+++ pkg/binder/binding/mock/binder_mock.go
@@ -55,3 +55,17 @@ func (mr *MockInterfaceMockRecorder) Bind(ctx, task, host, bindRequest any) *gom
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Bind", reflect.TypeOf((*MockInterface)(nil).Bind), ctx, task, host, bindRequest)
 }
+
+// Rollback mocks base method.
+func (m *MockInterface) Rollback(ctx context.Context, task *v1.Pod, host *v1.Node, bindRequest *v1alpha2.BindRequest) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "Rollback", ctx, task, host, bindRequest)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// Rollback indicates an expected call of Rollback.
+func (mr *MockInterfaceMockRecorder) Rollback(ctx, task, host, bindRequest any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Rollback", reflect.TypeOf((*MockInterface)(nil).Rollback), ctx, task, host, bindRequest)
+}