+++ pkg/binder/plugins/interface.go
@@ -19,4 +19,6 @@ type Plugin interface {
 		state *state.BindingState) error
 	PostBind(ctx context.Context, pod *v1.Pod, node *v1.Node, bindRequest *v1alpha2.BindRequest,
 		state *state.BindingState)
+	Rollback(ctx context.Context, pod *v1.Pod, node *v1.Node, bindRequest *v1alpha2.BindRequest,
+		state *state.BindingState) error
 }