+++ pkg/common/constants/constants.go
@@ -31,15 +31,17 @@ const (
 	TopOwnerMetadataKey = "kai.scheduler/top-owner-metadata"
 
 	// Annotations
-	PodGroupAnnotationForPod = "pod-group-name"
-	GpuFraction              = "gpu-fraction"
-	GpuFractionContainerName = "gpu-fraction-container-name"
-	GpuMemory                = "gpu-memory"
-	ReceivedResourceType     = "received-resource-type"
-	GpuFractionsNumDevices   = "gpu-fraction-num-devices"
-	MpsAnnotation            = "mps"
-	StalePodgroupTimeStamp   = "kai.scheduler/stale-podgroup-timestamp"
-	LastStartTimeStamp       = "kai.scheduler/last-start-timestamp"
+	PodGroupAnnotationForPod      = "pod-group-name"
+	GpuFraction                   = "gpu-fraction"
+	GpuFractionContainerName      = "gpu-fraction-container-name"
+	GpuMemory                     = "gpu-memory"
+	ReceivedResourceType          = "received-resource-type"
+	GpuFractionsNumDevices        = "gpu-fraction-num-devices"
+	MpsAnnotation                 = "mps"
+	StalePodgroupTimeStamp        = "kai.scheduler/stale-podgroup-timestamp"
+	LastStartTimeStamp            = "kai.scheduler/last-start-timestamp"
+	GpuSharingConfigMapAnnotation = "runai/shared-gpu-configmap"
+	NvidiaVisibleDevices          = "NVIDIA_VISIBLE_DEVICES"
 
 	// UsageDB Prometheus Selector
 	DefaultAccountingLabelKey   = "kai.scheduler/accounting"