+++ test/e2e/suites/allocate/node_order/placement_strategy_test.go
@@ -31,8 +31,6 @@ const (
 	binPackingPluginName = "binpack"
 	SpreadingPluginName  = "spread"
 	DefaultPluginName    = binPackingPluginName
-
-	gpuIndexEnvVarName = "NVIDIA_VISIBLE_DEVICES"
 )
 
 var _ = Describe("Placement strategy", Label(labels.Operated), Ordered, func() {
@@ -262,7 +260,7 @@ func getGpuIndexOfScheduledFractionPod(ctx context.Context, testCtx *testcontext
 	pod *v1.Pod) (string, error) {
 	for _, container := range pod.Spec.Containers {
 		for _, env := range container.Env {
-			if env.Name == gpuIndexEnvVarName {
+			if env.Name == constants.NvidiaVisibleDevices {
 				configMapName := env.ValueFrom.ConfigMapKeyRef.Name
 				wait.ForConfigMapCreation(ctx, testCtx.ControllerClient, pod.Namespace, configMapName)
 				cm, err := testCtx.KubeClientset.CoreV1().ConfigMaps(pod.Namespace).
@@ -274,5 +272,5 @@ func getGpuIndexOfScheduledFractionPod(ctx context.Context, testCtx *testcontext
 			}
 		}
 	}
-	return "", fmt.Errorf("could not find %s env var in pod", gpuIndexEnvVarName)
+	return "", fmt.Errorf("could not find %s env var in pod", constants.NvidiaVisibleDevices)
 }