+++ test/e2e/suites/allocate/resources/resources_test.go
@@ -6,11 +6,14 @@ package resources
 
 import (
 	"context"
+	"sync"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/watch"
 
 	v2 "github.com/NVIDIA/KAI-scheduler/pkg/apis/scheduling/v2"
 	"github.com/NVIDIA/KAI-scheduler/pkg/common/constants"
@@ -21,6 +24,7 @@ import (
 	"github.com/NVIDIA/KAI-scheduler/test/e2e/modules/resources/rd/queue"
 	"github.com/NVIDIA/KAI-scheduler/test/e2e/modules/utils"
 	"github.com/NVIDIA/KAI-scheduler/test/e2e/modules/wait"
+	runtimeClient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 var _ = Describe("Schedule pod with resource request", Ordered, func() {
@@ -85,6 +89,112 @@ var _ = Describe("Schedule pod with resource request", Ordered, func() {
 			Expect(err).NotTo(HaveOccurred())
 
 			wait.ForPodReady(ctx, testCtx.ControllerClient, pod)
+		})
+
+		It("Fraction GPU request - fill all the GPUs", Label(labels.ReservationPod), func(ctx context.Context) {
+			resources, err := capacity.GetClusterAllocatableResources(testCtx.KubeClientset)
+			Expect(err).NotTo(HaveOccurred())
+			numGPUs := int(resources.Gpu.Value())
+
+			numPods := numGPUs * 2
+			pods := make([]*v1.Pod, numPods)
+			for i := range numPods {
+				pods[i] = rd.CreatePodObject(testCtx.Queues[0], v1.ResourceRequirements{})
+				pods[i].Annotations = map[string]string{
+					constants.GpuFraction: "0.7",
+				}
+			}
+			errs := make(chan error, len(pods))
+			var wg sync.WaitGroup
+			for _, pod := range pods {
+				wg.Add(1)
+				go func() {
+					defer wg.Done()
+					_, err := rd.CreatePod(ctx, testCtx.KubeClientset, pod)
+					errs <- err
+				}()
+			}
+			wg.Wait()
+			close(errs)
+			for err := range errs {
+				Expect(err).NotTo(HaveOccurred(), "Failed to create pod")
+			}
+
+			wait.ForAtLeastNPodCreation(ctx, testCtx.ControllerClient, metav1.LabelSelector{
+				MatchLabels: map[string]string{constants.AppLabelName: "engine-e2e"},
+			}, numPods)
+
+			labelSelector, err := metav1.LabelSelectorAsSelector(&metav1.LabelSelector{
+				MatchLabels: map[string]string{constants.AppLabelName: "engine-e2e"},
+				MatchExpressions: []metav1.LabelSelectorRequirement{
+					{
+						Key:      constants.GPUGroup,
+						Operator: metav1.LabelSelectorOpExists,
+					},
+				},
+			})
+			wait.ForPodsWithCondition(ctx, testCtx.ControllerClient, func(event watch.Event) bool {
+				podList, ok := event.Object.(*v1.PodList)
+				if !ok {
+					return false
+				}
+				scheduledWithGpuGroup := 0
+				for _, pod := range podList.Items {
+					if !rd.IsPodScheduled(&pod) {
+						continue
+					}
+					if _, ok := pod.Labels[constants.GPUGroup]; !ok {
+						continue
+					}
+					scheduledWithGpuGroup++
+				}
+				return scheduledWithGpuGroup == numGPUs
+			}, runtimeClient.InNamespace(testCtx.Queues[0].Namespace),
+				runtimeClient.MatchingLabelsSelector{Selector: labelSelector})
+
+			var allPods v1.PodList
+			Expect(testCtx.ControllerClient.List(ctx, &allPods,
+				runtimeClient.InNamespace(testCtx.Queues[0].Namespace),
+				runtimeClient.MatchingLabelsSelector{Selector: labelSelector},
+			)).To(Succeed(), "Failed to list pods")
+			Expect(err).NotTo(HaveOccurred())
+			Expect(len(allPods.Items)).To(BeNumerically("==", numGPUs), "Expected exactly %d allocated pods, got %d", numGPUs, len(allPods.Items))
+
+			var allocatedPods []*v1.Pod
+			var configMaps v1.ConfigMapList
+			Expect(testCtx.ControllerClient.List(ctx, &configMaps,
+				runtimeClient.InNamespace(testCtx.Queues[0].Namespace),
+			)).To(Succeed(), "Failed to list config maps")
+
+			gpuGroupsMap := make(map[string][]*v1.Pod)
+			cmByGPUIndex := make(map[string][]*v1.ConfigMap)
+			for _, pod := range allPods.Items {
+				Expect(rd.IsPodScheduled(&pod)).To(BeTrue(), "Expected pod to be scheduled", pod.Name, pod.Labels, pod.Status.Conditions)
+				allocatedPods = append(allocatedPods, &pod)
+
+				group, ok := pod.Labels[constants.GPUGroup]
+				Expect(ok).To(BeTrue(), "Expected GPU group label to be found on pod %s", pod.Name)
+				gpuGroupsMap[group] = append(gpuGroupsMap[group], &pod)
+
+				cmName := pod.Annotations[constants.GpuSharingConfigMapAnnotation]
+				for _, cm := range configMaps.Items {
+					if cm.Name != cmName {
+						continue
+					}
+					index := cm.Data[constants.NvidiaVisibleDevices]
+					cmByGPUIndex[index] = append(cmByGPUIndex[index], &cm)
+					break
+				}
+			}
+			Expect(len(allocatedPods)).To(BeNumerically("==", numGPUs), "Expected exactly %d allocated pods, got %d", numGPUs, len(allocatedPods))
+
+			for group, pods := range gpuGroupsMap {
+				Expect(len(pods)).To(Equal(1), "Expected one pod per group, got %d for group %s", len(pods), group)
+			}
+
+			for index, cm := range cmByGPUIndex {
+				Expect(len(cm)).To(Equal(1), "Expected one config map per gpu index, got %d for index %s", len(cm), index)
+			}
 		})
 
 		It("GPU memory request - valid", Label(labels.ReservationPod), func(ctx context.Context) {